backend/requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
audioread==3.1.0
bcrypt==4.1.3
beautifulsoup4==4.14.2
black==25.11.0
blinker==1.9.0
boto3==1.40.76
botocore==1.40.76
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
decorator==5.2.1
deepface==0.0.95
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.1
filelock==3.20.0
fire==0.7.1
flake8==7.3.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fsspec==2025.10.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
HyperPyYAML==1.2.2
idna==3.11
iniconfig==2.3.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jq==1.10.0
keras==3.12.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.45.1
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.4
motor==3.3.1
mpmath==1.3.0
msgpack==1.1.2
mtcnn==1.0.0
mypy==1.18.2
mypy_extensions==1.1.0
namex==0.1.0
networkx==3.5
numba==0.62.1
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
protobuf==6.33.1
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
PySocks==1.7.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.2.0
rsa==4.9.1
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
s3transfer==0.14.0
s5cmd==0.2.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
speechbrain==1.0.3
spotipy==2.25.1
starlette==0.37.2
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.25.0
watchfiles==1.1.1
Werkzeug==3.1.3
wrapt==2.0.1